+++ packages/nesd/lib/nes/cartridge/mapper/mapper_state.dart
@@ -4,6 +4,7 @@ import 'package:nesd/exception/unsupported_mapper.dart';
 import 'package:nesd/nes/cartridge/mapper/axrom_state.dart';
 import 'package:nesd/nes/cartridge/mapper/bandai_fcg_state.dart';
 import 'package:nesd/nes/cartridge/mapper/cnrom_state.dart';
+import 'package:nesd/nes/cartridge/mapper/color_dreams_state.dart';
 import 'package:nesd/nes/cartridge/mapper/fme7_state.dart';
 import 'package:nesd/nes/cartridge/mapper/gxrom_state.dart';
 import 'package:nesd/nes/cartridge/mapper/mapper176_state.dart';
@@ -43,6 +44,7 @@ abstract class MapperState {
       5 => MMC5State.deserialize(reader),
       7 => AXROMState.deserialize(reader),
       9 => MMC2State.deserialize(reader),
+      11 => ColorDreamsState.deserialize(reader),
       16 => BandaiFCGState.deserialize(reader),
       19 => Namco163State.deserialize(reader),
       21 || 22 || 23 || 25 => VRC24State.deserialize(reader, id),